pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medicine-assistant"
version = "0.1.0"
description = "An AI Agent that assists doctors in picking the correct medicine and dosage for patients"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ibrahim Younis"}
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.27",
    "langchain-chroma>=0.1.0",
    "langchain-huggingface>=0.1.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pypdf",
    "supabse",
    "psycopg2",
    "flask",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]

[project.scripts]
medicine-assistant = "main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
